feat: :recycle: refactor code for workshop
